:bug: Fix author sometimes not showing up under certain condition
Commit messages from GitLab sometimes (usually) have a `\n` newline appended to them. Depending on the speed of the program, it would've added the commit author to the first couple commits before having it stripped (as it strips newlines and anything after), due to the author being appended to the commit message. Just flipped the order - now it strips newlines then adds author.