:sparkles: :construction: :recycle: :alembic: Begin 'pluginization'
For Githook's future it wouldn't have been great if every event was stuck in app.py - it would've gotten messy over time and near unmaintainable. This aims to make Githook run in a 'modular' way so that all events are in its own file and run dynamically when needed (if they exist).

Big thanks to @byxor, without whom this would have been an absolute nightmare to figure out.